refactor(P5): add CI + clippy cleanup — correctness/suspicious as hard errors
- .github/workflows/ci.yml: build(release) → clippy(-D correctness -D
  suspicious) → cargo test (CPU only; real-model tests skip when files are
  missing).
- [workspace.lints.clippy] allow list (rationale in comments):
  too_many_arguments (kernel argument spec) · needless_range_loop (numeric
  mirror) · type_complexity · excessive_precision · approx_constant.
- Mechanical cleanup of 31 files via clippy --fix (both gates confirmed
  bit-identical).
- clippy caught 3 real defects, fixed:
  * the qwen4exp test's  — a dead assertion that always passed.
  * rawhip's unused y-reuse cache flag (reason preserved in a comment) · the
    q4acc use6 flag.
  * 3 doc comments left over from the P2 move in rawhip/mod.rs that had
    stuck to RawCtx.
- The 4 raw stream handle APIs in graph.rs made unsafe (call sites in unsafe
  blocks).
- 1 CI exception documented: arc_with_non_send_sync (VL single-threaded
  sharing only).

+++ crates/backend-gpu/src/rawhip/vit.rs
@@ -109,8 +109,8 @@ impl Vit {
         ];
         let gx = t.div_ceil(32) as u32;
         let gy = n_out.div_ceil(8) as u32;
-        let gz = gy.div_ceil(65535) as u32;
-        self.ctx.launch3("gemm_f32t", gx, gy.min(65535) as u32, gz, 256, &mut args)
+        let gz = gy.div_ceil(65535);
+        self.ctx.launch3("gemm_f32t", gx, gy.min(65535), gz, 256, &mut args)
     }
 
     /// toks: merge-major [t][n_embd] (conv+pos+bias 적용된 입력), yx: [t][2].
